mdbx: add `MDBX_xPL_PREALLOC_FOR_RADIXSORT` build options.

+++ src/internals.h
@@ -793,7 +793,7 @@ typedef struct MDBX_lockinfo {
 #define MAX_MAPSIZE MAX_MAPSIZE64
 #define MDBX_READERS_LIMIT                                                     \
   ((MAX_PAGESIZE - sizeof(MDBX_lockinfo)) / sizeof(MDBX_reader))
-#define MDBX_PGL_LIMIT MAX_PAGENO
+#define MDBX_PGL_LIMIT ((size_t)MAX_PAGENO)
 #else
 #define MDBX_READERS_LIMIT 1024
 #define MAX_MAPSIZE MAX_MAPSIZE32
